+++ CHANGELOG.md
@@ -11,7 +11,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ### Added 
 
-- A new Directive `@extends` that enables types to inherit attributes from other types.
+- A new Directive `@inherits` that enables types to inherit attributes from other types.
 
 ## v5.15.1